Break out validation helpers so they can be tested
